[WIP] don't benchmark exceptions...
