+++ Project.toml
@@ -15,3 +15,6 @@ HerbGrammar = "1"
 HerbSpecification = "1"
 RuntimeGeneratedFunctions = "0.5.16"
 julia = "1.10"
+
+[workspace]
+projects = ["test", "benchmark"]
